+++ src/ext/cmdhdl.py
@@ -1,168 +1,119 @@
-import discord
+"""
+Message and command handling discord.py extension.
+
+Command processing, command invoking, error handling, and partial command support here.
+"""
 import json
+import warnings
 import traceback
-import sys
 import asyncio
+import discord
 from discord.ext import commands
-from ext.const import SETFILE, LASTWRDFILE, STRFILE, fix_settings, get_prefix, cmdHdlLogger, log
-import special
 from ext import excepts
+from ext.const import SETFILE, LASTWRDFILE, STRFILE, get_prefix, cmdHdlLogger
 from modules.chat import chat
-from datetime import datetime
 from modules.consolemod import style
+from modules.tools import wrdssep
+import merlin
+import special
+BOT = None
 
 
-async def init(bot: commands.Bot):
-    # lastword
-    try:
-        json.load(open(LASTWRDFILE, 'r'))
-    except json.JSONDecodeError:
-        open(LASTWRDFILE, 'w').write("{}")
-
-
-async def fix_set(bot: commands.Bot, message: discord.Message):
+async def get_cur_prefix(message):
     settings = json.load(open(SETFILE, 'r'))
-    try:
-        # test if entry presents
-        settings["cmdHdl"]['improveExp']
-    except Exception:
-        fix_settings(message.guild)  # fix guild settings
+    prefix = e = None
+    prefixes = settings[f"g{message.guild.id}"]["prefix"]
+    for p in prefixes:
+        if message.content.startswith(p):
+            prefix = p
+            break
+    return prefix
 
 
-async def proc_cmd(bot: commands.Bot, message: discord.Message):
-    settings = json.load(open(SETFILE, 'r'))
-    if message.content.startswith(get_prefix(bot, message)) and message.channel.name != 'merlin-chat':
-        msgtoSend = f'{message.author} has issued command: '
-        print(msgtoSend + style.green + message.content + style.reset)
-        cmdHdlLogger.info(msgtoSend + message.content)
+async def proc_cmd(message: discord.Message):
+    """
+    Process commands.
+
+    Check if it is a valid command, then
+    log and invoke it.
+    """
+    if message.channel.name == 'merlin-chat':
+        return
+    prefix = None
+    prefixes = await BOT.get_prefix(message)
+    for p in prefixes:
+        if message.content.startswith(p):
+            prefix = p
+            break
+    if prefix is None:
+        return  # not a cmd
+    with warnings.catch_warnings(record=True) as w:
+        cmd = BOT.get_command(message.content[len(prefix):], last_gud=True)
+        if len(w) == 1:
+            # if issubclass(w[-1].category, excepts.AmbiguousSearchQuery):
+            #     return await message.channel.send(w[-1].message)
+            await message.channel.send(w[-1].message)
+        if cmd is None:
+            return  # not a cmd
+        cmdHdlLogger.info(f'{message.author} has issued command: {message.content}')
         try:
-            await log(message.channel.mention + ' ' + msgtoSend + '`' + message.content + '`', guild=message.channel.guild)
-        except AttributeError:
+            await BOT.netLogger(f"{message.channel.mention} {message.author} has issued command: `{message.content}`", guild=message.guild)
+        except AttributeError:  # DM
             pass
         try:
-            await bot.process_commands(message)
-            if settings[f'g{message.guild.id}']["cmdHdl"]["delIssue"]:
+            ctx = await BOT.get_context(message)
+            await BOT.invoke(ctx)
+            if BOT.db['sets'][f'g{message.guild.id}']["cmdHdl"]["delIssue"]:
                 await message.delete()
             return 0
-        except discord.ext.commands.errors.CommandNotFound:
-            pass
-        except discord.errors.NotFound:
+        except (discord.ext.commands.errors.CommandNotFound, discord.errors.NotFound):
             pass
         except Exception:
-            await message.channel.send(f'{message.author.mention}, there was an error trying to execute that command! :(')
-            print(traceback.format_exc())
+            await message.add_reaction("❌")
+            cmdHdlLogger.debug(traceback.format_exc(limit=5))
 
 
-async def save_quote(bot: commands.Bot, message: discord.Message):
-    lastword = json.load(open(LASTWRDFILE, 'r'))
+async def save_quote(bot: merlin.Bot, message: discord.Message):
+    lastword = bot.db['lastwrds']
     try:
         lastword[f'g{message.guild.id}'][str(message.author.id)] = message.id
     except KeyError:
         lastword[f'g{message.guild.id}'] = {message.author.id: message.id}
-    json.dump(lastword, open(LASTWRDFILE, 'w'))
 
 
-async def chat_hdl(bot: commands.Bot, message: discord.Message):
-    settings = json.load(open(SETFILE, 'r'))
+async def chat_hdl(bot: merlin.Bot, message: discord.Message):
+    settings = bot.db['sets']
     if not isinstance(message.channel, discord.DMChannel) and message.channel.name == 'merlin-chat' and not message.author.bot:
         res, chatbot = await chat.response(message)
         await message.channel.send(res)
     elif not isinstance(message.channel, discord.DMChannel) and not message.author.bot and settings[f'g{message.guild.id}']["cmdHdl"]["improveExp"]:
         msgs = await message.channel.history(limit=2).flatten()
-        await asyncio.gather(chat.save(message.content, msgs[1].content))   
+        await asyncio.gather(chat.save(message.content, msgs[1].content))
 
 
 # discord extension
-def setup(bot: commands.Bot):
-    bot.loop.create_task(init(bot))  # init chk
+def setup(bot: merlin.Bot):
+    """Ext setup."""
+    global BOT
+    BOT = bot
 
     @bot.event
     async def on_message(message: discord.Message):
         # run pre-cmd hooks
         if await special.pre_on_message(message):
             return 0
-        # check if settings works
-        await fix_set(bot, message)
         # gather and run in parallel (nowait)
-        await asyncio.gather(save_quote(bot, message), proc_cmd(bot, message), chat_hdl(bot, message))
+        await asyncio.gather(save_quote(bot, message), proc_cmd(message), chat_hdl(bot, message))
         await special.post_on_message(message)   # set fn for callback when done
 
     @bot.event
-    async def on_command_error(ctx, error):
-        settings = json.load(open(SETFILE, 'r'))
-        stringTable = json.load(open(STRFILE, 'r'))
-        try:
-            raise error
-        except Exception:
-            # This tells the issuer that the command cannot be used in DM
-            if isinstance(error, commands.errors.NoPrivateMessage):
-                try:
-                    bot.loop.create_task(ctx.author.send(f':x::lock: {ctx.command} cannot be used in Private Messages.'))
-                    return 3
-                except discord.HTTPException:
-                    return 3
-            # This prevents any commands with local handlers being handled here in on_command_error.
-
-            # This prevents any cogs with an overwritten cog_command_error being handled here.
-            if ctx.cog and ctx.cog._get_overridden_method(ctx.cog.cog_command_error) is not None:
-                return
-
-            # Anything in ignored will return and prevent anything happening.
-            if isinstance(error, commands.errors.CommandNotFound):
-                try:
-                    if settings[f'g{ctx.guild.id}']['cmdHdl']['cmdNotFound']:
-                        bot.loop.create_task(
-                            ctx.send(":interrobang: Welp, I've no idea. Command not found!"))
-                except KeyError:
-                    bot.loop.create_task(ctx.send(
-                        ":interrobang: :two: :x:\n<:err:740034702743830549> Command not found!\n"
-                        "<:warn:739838316374917171> something went wrong, please run `/settings`"
-                    ))
-                return 2
-            if isinstance(error, commands.MissingRequiredArgument) or isinstance(error, commands.BadArgument):
-                return bot.loop.create_task(ctx.invoke(bot.get_command('help'), cmdName=ctx.command.qualified_name))
-
-            if isinstance(error, commands.errors.DisabledCommand):
-                return bot.loop.create_task(ctx.send(embed=discord.Embed(
-                    title=f':no_entry: {ctx.command} has been disabled.',
-                    description=f':x: `{ctx.message.content}`',
-                    color=0xff0000
-                )))
-
-            if isinstance(error, excepts.NotMod):
-                bot.loop.create_task(ctx.send(
-                    f"`{ctx.author} is not in the sudoers file.  This incident will be reported.`"))
-
-            if hasattr(ctx.command, 'on_error'):
-                return
-
-            if isinstance(error, commands.errors.CommandInvokeError):
-                bot.loop.create_task(ctx.send(embed=discord.Embed(
-                    title='<:err:740034702743830549> uh oh. An exception has occurred during the execution of the command',
-                    description=stringTable['CommandInvokeError'].format(
-                        content=ctx.message.content),
-                    timestamp=datetime.utcnow(),
-                    color=0xff0000
-                )))
-
-            if isinstance(error, commands.errors.NotOwner):
-                bot.loop.create_task(ctx.send(stringTable['notOwner']))
-                return 6
-            if isinstance(error, commands.errors.ConversionError):
-                bot.loop.create_task(ctx.send(
-                    ':bangbang: Hey bud, seems like you tried to input some invalid type of arguments to the command call!\n'
-                    'Either CoNsUlT a PsYcHiAtRiSt or check the usage. Please!'))
-                return 4
-
-            if isinstance(error, commands.errors.BadArgument):
-                bot.loop.create_task(ctx.send(
-                    ':grey_question: Whoops. The discord special expression you have specified when issuing that command is invalid.'
-                    ':mag: This error occurrs usually because of the bot fails to find the object.'))
-                return 4
-            if isinstance(error, excepts.NotMod):
-                return await ctx.send(str(error))
-
-            # All other Errors not returned come here. And we can just print the default TraceBack.
-            print(f"{style.red2}Ignoring exception in command {ctx.message.content}:\n{style.grey}{traceback.format_exc()}{style.reset}", file=sys.stderr)
-            await log(f'Ignoring exception in command `{ctx.message.content}`:\n\n```{str(traceback.format_exc())}\n```', guild=ctx.guild)
-            return 1
+    async def on_command_error(ctx, e):
+        # This prevents any commands with local handlers being handled here in on_command_error.
+        if hasattr(ctx.command, 'on_error'):
+            return
+        # This prevents any cogs with an overwritten cog_command_error being handled here.
+        if ctx.cog and ctx.cog._get_overridden_method(ctx.cog.cog_command_error) is not None:
+            return
+        err_code = await BOT.errhdl_g(ctx, e)
+        if err_code:
+            await BOT.netLogger(f"FAIL {err_code}: `{ctx.message.content}`", ctx.guild)